Make the menu appear when introHeight is large
To that effect, introduce a new hero is-bottom-menu class which uses
viewport height at 80% to give room for the menu (can't be perfect since
we of course can't mix vh and rem units). Also make sure we use that
class instead of is-large when introHeight is set to large.

That should match the behavior discussed previously.