feat: Put time format and timezone in params instead of params.home
Currently, the `timeZone` and `timeFormat` parameters are in the `params.home` section of the configuration file. This is not consistent with the `dateFormat` that is directly in the `params` section and I don't see why this setting should be just for the home page.

The advantage of having the settings in the `params` section is with multi-lingual sites : you can have "HH:mm" for European version and "h:mm A" for US version of the page. If the setting is in `params.home` the whole `params.home` section has to be duplicated. Indeed, if you have a `language.xx.params.home` section, it does not *override* the `params.home` with the new settings, but it *replaces* the whole section.

This change keeps the compatibility with the version where the settings are in home.