feat: add `timeUpdatePeriod` for periodic localtime update
timeUpdatePeriod is a new setting (in params.home) that controls periodical update of the local time. If set to zero, the localtime is not updated. If set to a value greater than zero, the localtime is updated every `timeUpdatePeriod` milliseconds. Typical values for this parameter are either 0 (update disabled) or 1000 (update every second).